+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 
 ### Added
 
-- None
+- **[BC Break]** `MemberContainer::remove` methods get added to the interface.
 
 ### Fixed
 
@@ -22,6 +22,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 ### Removed
 
 - **[BC Break]** `OrderedList` is removed use `OuterList`.
+- **[BC Break]** `MemberList::remove` and `MemberOrderedMap::remove` methods are removed from the interface.
 
 ## [0.7.0] - 2023-02-06